+++ moor_ffi/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.6.0
+
+- Added `moor_contains` sql function to support case-sensitive contains
+
 ## 0.5.0
 
 - Provide mathematical functions in sql (`pow`, `power`, `sin`, `cos`, `tan`, `asin`, `atan`, `acos`, `sqrt`)